Move helpers to _message.py
